Move App/Root to components
