--- .babelrc
@@ -1,3 +1,0 @@
-module.exports = {
-    "presets": ["babel-preset-react-app"]
-}
