+++ babel.config.js
@@ -1,0 +1,3 @@
+module.exports = process.env.CYPRESS_ENV ? {} : {
+    "presets": ["babel-preset-react-app"]
+}
